Remove nvme multipath default off patch (#7115)
NVME multipath was merged in cm2 in 0af13f0 to the 5.15 kernel, but the default behavior was patched to off to ensure current customers were not impacted. Now for 6.6, remove the patch to restore default upstream behavior of nvme multipath on. Note that nvme_multipath can still be turned off by setting nvme_core.multipath=n on the commandline.